requirements.txt
flask
flask-cors
requests
openai
flask-socketio
python-dateutil
python-dotenv
psycopg2-binary
